+++ collective/teaser/config.py
@@ -1,7 +1,7 @@
 PROJECTNAME = "collective.teaser"
 DEFAULT_IMPORTANCE = '3',
 ADD_PERMISSIONS = {
-    "Teaser"        : "collective.teaser: Add Teaser",
+    "Teaser": "collective.teaser: Add Teaser",
 }
 
 from Products.CMFCore.permissions import setDefaultRoles