+++ collective/teaser/vocabularies.py
@@ -1,8 +1,10 @@
 from zope.interface import directlyProvides
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleVocabulary
-from collective.teaser import MsgFact as _
+from zope.i18nmessageid import MessageFactory
 from plone.app.imaging.utils import getAllowedSizes
+
+_ = MessageFactory('collective.teaser')
 
 
 def ImportanceVocabulary(context):