+++ docs/HISTORY.rst
@@ -1,10 +1,11 @@
+
 Changelog
 =========
 
 2.2dev
 ------
 
-- Add display_columns for teaser portlet.
+- Add ``display_columns`` for teaser portlet.
   [rnix, 2012-08-16]
 
 - Add portlet header for teaser portlet.